* adjusted online page to be a little less wordy * fixed some issues with setPref.  should be good to go.
git-svn-id: http://abadthing.com/tmbo/trunk@148 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43

+++ web/offensive/setPref.php
@@ -12,10 +12,10 @@
 	require_once('offensive/assets/getPrefs.inc');
 	require_once('offensive/assets/functions.inc');
 
-	$prefid = sqlEscape( $_REQUEST['p'] );
-	$valueid = sqlEscape( $_REQUEST['v'] );
+	$prefid = sqlEscape( array_key_exists("p", $_REQUEST) ? $_REQUEST['p'] : "");
+	$valueid = sqlEscape( array_key_exists("v", $_REQUEST) ?$_REQUEST['v'] : "");
 
-	if( $prefid ) {
+	if( strlen($prefid) > 0 ) {
 	
 		$sql = "SELECT user_preferences.id FROM user_preferences WHERE nameid=$prefid AND userid = " . $_SESSION['userid'];	
 		
@@ -27,13 +27,11 @@
 			if( $valueid != "" ) {
 				// if we have a value and an existing preference, update it.
 				$sql = "UPDATE user_preferences SET valueid=$valueid WHERE id=$rowid";
-			}
-			else {
+			} else {
 				// if we have an existing value but no new value, delete the existing record
 				$sql = "DELETE FROM user_preferences WHERE id=$rowid";
 			}
-		}
-		else {
+		} else {
 			// if no preference for this pref name and user exists, add it.
 			$sql = "INSERT INTO user_preferences (userid, nameid, valueid) VALUES ( " . $_SESSION['userid'] . ", $prefid, $valueid )";
 		}
@@ -44,7 +42,7 @@
 	
 	}
 	
-	if( is_numeric( $_REQUEST['sq'] ) ) {
+	if( array_key_exists("sq", $_REQUEST) && is_numeric( $_REQUEST['sq'] ) ) {
 		$squelch = sqlEscape( $_REQUEST['sq'] );
 		$sql = "insert into offensive_squelch (userid, squelched) VALUES ( " . $_SESSION['userid'] . ", $squelch )";
 		tmbo_query( $sql );
@@ -52,7 +50,7 @@
 		$_SESSION['prefs'] = $prefs;
 	}
 
-	if( is_numeric( $_REQUEST['unsq'] ) ) {
+	if( array_key_exists("unsq", $_REQUEST) && is_numeric( $_REQUEST['unsq'] ) ) {
 		$squelch = sqlEscape( $_REQUEST['unsq'] );
 		$userid = $_SESSION['userid'];
 		$sql = "delete from offensive_squelch where userid=$userid AND squelched=$squelch";
@@ -62,6 +60,10 @@
 	}
 
 
-	header( "Location: " . $_SERVER['HTTP_REFERER'] );
+	if(array_key_exists("HTTP_REFERER", $_SERVER)) {
+		header( "Location: " . $_SERVER['HTTP_REFERER'] );
+	} else {
+		echo "<html><head><script type=\"text/javascript\">history.go(-1);</script></head><body /></html>";
+	}
 
 ?>
